+++ ratesAndConversions/src/main/kotlin/aze/talmir/task/ratesconversions/ui/main/RatesConversionsViewModel.kt
@@ -2,6 +2,7 @@ package aze.talmir.task.ratesconversions.ui.main
 
 import androidx.lifecycle.ViewModel
 import aze.talmir.task.ratesconversions.data.IRatesConversionsRepository
+import java.math.BigDecimal
 
 /**
  * Classic ViewModel class. Helps UI (fields) to communicate with
@@ -11,6 +12,6 @@ class RatesConversionsViewModel(
     private val ratesConversionsRepository: IRatesConversionsRepository
 ) : ViewModel() {
 
-    suspend fun getRatesAndConversions(baseCurrency: String, coefficient: Double) =
+    suspend fun getRatesAndConversions(baseCurrency: String, coefficient: BigDecimal) =
         ratesConversionsRepository.getRates(baseCurrency, coefficient)
 }
